refactor(og): share frame across OG and Twitter image routes
Extract renderOgImage()/renderNotFoundImage() into src/utils/og-image.tsx
so the four slug image routes share one layout and branding. Each route now
delegates to a per-domain og-image-content builder, with the OG vs Twitter
difference reduced to a variant flag controlling the footer identity line
(site host vs Twitter handle).
